Move over some unit tests
